refactor: improve typings of addon-storyshots-puppeteer
