+++ addons/storyshots/storyshots-puppeteer/package.json
@@ -22,21 +22,25 @@
     "README.md"
   ],
   "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "scripts": {
     "prepare": "node ../../../scripts/prepare.js"
   },
   "dependencies": {
     "@storybook/node-logger": "5.3.0-alpha.13",
     "@storybook/router": "5.3.0-alpha.13",
+    "@types/jest-image-snapshot": "^2.8.0",
     "core-js": "^3.0.1",
     "jest-image-snapshot": "^2.8.2",
     "regenerator-runtime": "^0.13.3"
   },
   "peerDependencies": {
-    "@storybook/addon-storyshots": "5.2.0-beta.13",
+    "@storybook/addon-storyshots": "5.3.0-alpha.13",
+    "@types/puppeteer": "^1.19.0",
     "puppeteer": "^1.12.2"
   },
   "optionalDependencies": {
+    "@types/puppeteer": "^1.19.0",
     "puppeteer": "^1.12.2"
   },
   "publishConfig": {
