+++ addons/storyshots/storyshots-puppeteer/tsconfig.json
@@ -2,7 +2,8 @@
   "extends": "../../../tsconfig.json",
   "compilerOptions": {
     "rootDir": "./src",
-    "types": ["webpack-env", "node"]
+    "types": ["webpack-env", "node"],
+    "declaration": true
   },
   "include": [
     "src/**/*"
